Refactor indexer API structure
