--- crates/indexer/src/api/dto/mod.rs
@@ -1,9 +1,0 @@
-pub mod factories;
-pub mod offer_participants;
-pub mod offer_utxo;
-pub mod offers;
-
-pub use factories::*;
-pub use offer_participants::*;
-pub use offer_utxo::*;
-pub use offers::*;
