--- crates/indexer/src/api/handlers/mod.rs
@@ -1,9 +1,0 @@
-mod factories;
-mod offer_participants;
-mod offer_utxo;
-mod offers;
-
-pub use factories::*;
-pub use offer_participants::*;
-pub use offer_utxo::*;
-pub use offers::*;
