+++ crates/indexer/src/api/mod.rs
@@ -1,11 +1,11 @@
-mod db;
-mod dto;
 mod error;
-mod handlers;
+mod factories;
+mod offers;
 mod params;
 pub mod server;
+mod state;
 pub mod utils;
 
 pub use error::*;
 pub use params::*;
-pub use server::*;
+pub use state::AppState;
