+++ crates/indexer/src/api/offers/db.rs
@@ -5,13 +5,13 @@ use sqlx::{PgPool, Postgres, QueryBuilder};
 use uuid::Uuid;
 
 use crate::api::OfferFilters;
-use crate::api::dto::{
-    FactoryByScriptResponse, FactoryByScriptRow, OfferDetailsResponse, OfferListItemFull,
-    OfferListItemShort, OfferUtxoDto, ParticipantDto,
-};
 use crate::models::{
-    FactoryStatus, OfferModel, OfferModelShort, OfferParticipantModel, OfferStatus, OfferUtxoModel,
+    OfferModel, OfferModelShort, OfferParticipantModel, OfferStatus, OfferUtxoModel,
     ParticipantType, UtxoType,
+};
+
+use super::dto::{
+    OfferDetailsResponse, OfferListItemFull, OfferListItemShort, OfferUtxoDto, ParticipantDto,
 };
 
 #[tracing::instrument(
@@ -24,7 +24,7 @@ use crate::models::{
         asset = filters.asset
     )
 )]
-pub async fn fetch_offers_full_info_filtered(
+pub async fn fetch_full_info_filtered(
     db: &PgPool,
     filters: OfferFilters,
 ) -> Result<Vec<OfferListItemFull>, sqlx::Error> {
@@ -80,7 +80,7 @@ pub async fn fetch_offers_full_info_filtered(
         asset = filters.asset
     )
 )]
-pub async fn fetch_offers_short_info_filtered(
+pub async fn fetch_short_info_filtered(
     db: &PgPool,
     filters: OfferFilters,
 ) -> Result<Vec<OfferListItemShort>, sqlx::Error> {
@@ -130,7 +130,7 @@ pub async fn fetch_offers_short_info_filtered(
     skip(db, offer_id),
     fields(%offer_id)
 )]
-pub async fn fetch_offer_full_info_by_id(
+pub async fn fetch_full_info_by_id(
     db: &PgPool,
     offer_id: Uuid,
 ) -> Result<Option<OfferListItemFull>, sqlx::Error> {
@@ -170,7 +170,7 @@ pub async fn fetch_offer_full_info_by_id(
         ids_count = %ids.len(),
     )
 )]
-pub async fn fetch_offer_details_by_ids(
+pub async fn fetch_details_by_ids(
     db: &PgPool,
     ids: &[Uuid],
 ) -> Result<Vec<OfferDetailsResponse>, sqlx::Error> {
@@ -238,7 +238,7 @@ pub async fn fetch_offer_details_by_ids(
     skip(db, offer_id),
     fields(%offer_id)
 )]
-pub async fn fetch_offer_participants_history(
+pub async fn fetch_participants_history(
     db: &PgPool,
     offer_id: Uuid,
 ) -> Result<Vec<ParticipantDto>, sqlx::Error> {
@@ -303,63 +303,11 @@ pub async fn fetch_latest_participants(
 }
 
 #[tracing::instrument(
-    name = "Fetching factories by script from DB",
-    skip(db, script_pubkey),
-    fields(script_pubkey = %script_pubkey.to_hex())
-)]
-pub async fn fetch_factories_by_script(
-    db: &PgPool,
-    script_pubkey: &[u8],
-) -> Result<Vec<FactoryByScriptResponse>, sqlx::Error> {
-    let rows = sqlx::query_as!(
-        FactoryByScriptRow,
-        r#"
-        SELECT DISTINCT ON (factory.id)
-            factory.id,
-            factory.factory_asset_id,
-            factory.program_script_pubkey,
-            factory.current_status AS "current_status: FactoryStatus",
-            factory.issuing_utxos_count,
-            factory.reissuance_flags,
-            factory.created_at_height,
-            factory.created_at_txid,
-            factory_auth.txid AS auth_txid,
-            factory_auth.vout AS auth_vout,
-            factory_auth.script_pubkey AS auth_script_pubkey,
-            factory_auth.created_at_height AS auth_created_at_height,
-            factory_program_utxo.txid AS program_txid,
-            factory_program_utxo.vout AS program_vout,
-            factory_program_utxo.created_at_height AS program_created_at_height
-        FROM factory_auths AS factory_auth
-        INNER JOIN factories AS factory
-            ON factory.id = factory_auth.factory_id
-        INNER JOIN factory_utxos AS factory_program_utxo
-            ON factory_program_utxo.factory_id = factory.id
-           AND factory_program_utxo.spent_txid IS NULL
-        WHERE factory_auth.script_pubkey = $1
-          AND factory_auth.spent_txid IS NULL
-          AND factory.current_status = 'active'
-        ORDER BY factory.id, factory_auth.created_at_height DESC
-        "#,
-        script_pubkey,
-    )
-    .fetch_all(db)
-    .await?;
-
-    let factories = rows
-        .into_iter()
-        .map(FactoryByScriptResponse::from)
-        .collect();
-
-    Ok(factories)
-}
-
-#[tracing::instrument(
     name = "Fetching offer ids by script from DB",
     skip(db, script_pubkey),
     fields(script_pubkey = %script_pubkey.to_hex())
 )]
-pub async fn fetch_offer_ids_by_script(
+pub async fn fetch_ids_by_script(
     db: &PgPool,
     script_pubkey: &[u8],
 ) -> Result<Vec<Uuid>, sqlx::Error> {
@@ -385,7 +333,7 @@ pub async fn fetch_offer_ids_by_script(
     skip(db, offer_id),
     fields(%offer_id)
 )]
-pub async fn fetch_offer_utxos_history(
+pub async fn fetch_utxos_history(
     db: &PgPool,
     offer_id: Uuid,
 ) -> Result<Vec<OfferUtxoDto>, sqlx::Error> {
