+++ crates/indexer/src/api/offers/mod.rs
@@ -1,0 +1,6 @@
+mod db;
+mod dto;
+mod handlers;
+mod routes;
+
+pub use routes::routes;
