+++ crates/indexer/src/api/server.rs
@@ -1,43 +1,21 @@
 use std::sync::Arc;
 
-use axum::{
-    Router,
-    routing::{get, post},
-};
+use axum::Router;
 use sqlx::PgPool;
 use tokio::net::TcpListener;
 use tower_http::request_id::{self, MakeRequestUuid, RequestId};
 use tower_http::trace::TraceLayer;
 
-use crate::api::handlers::{
-    get_factories_by_script, get_latest_offer_participants, get_offer_details,
-    get_offer_details_batch, get_offer_ids_by_script, get_offer_participants_history,
-    get_offer_utxos_history, get_offers_full_info, get_offers_short_info,
-};
-
-pub struct AppState {
-    pub db: PgPool,
-}
+use crate::api::factories;
+use crate::api::offers;
+use crate::api::state::AppState;
 
 pub async fn run_server(listener: TcpListener, db_pool: PgPool) {
     let state = Arc::new(AppState { db: db_pool });
 
     let app = Router::new()
-        .route("/factories/by-script", get(get_factories_by_script))
-        .route("/offers", get(get_offers_short_info))
-        .route("/offers/full", get(get_offers_full_info))
-        .route("/offers/batch", post(get_offer_details_batch))
-        .route("/offers/by-script", get(get_offer_ids_by_script))
-        .route("/offers/{id}", get(get_offer_details))
-        .route(
-            "/offers/{id}/participants/history",
-            get(get_offer_participants_history),
-        )
-        .route(
-            "/offers/{id}/participants",
-            get(get_latest_offer_participants),
-        )
-        .route("/offers/{id}/utxos", get(get_offer_utxos_history))
+        .merge(factories::routes())
+        .merge(offers::routes())
         .with_state(state)
         .layer(
             TraceLayer::new_for_http().make_span_with(|request: &axum::http::Request<_>| {
