+++ crates/indexer/src/api/state.rs
@@ -1,0 +1,5 @@
+use sqlx::PgPool;
+
+pub struct AppState {
+    pub db: PgPool,
+}
